+++ examples/log.rs
@@ -1,11 +1,11 @@
+extern crate futures;
+
 extern crate slog_loggly;
 
 #[macro_use]
 extern crate slog;
 
-use std::thread;
-
-use std::time::Duration;
+use futures::Future;
 
 use slog_loggly::LogglyDrain;
 
@@ -17,19 +17,18 @@ fn main() {
     let loggly_tag = "some-app";
 
     // Create a custom Loggly drain.
-    let drain = LogglyDrain::builder(loggly_token, loggly_tag)
+    let (drain, mut fhandle) = LogglyDrain::builder(loggly_token, loggly_tag)
         .debug_mode(true)
-        .spawn_thread()
-        .fuse();
+        .spawn_thread();
 
     // Create a logger.
-    let logger = Logger::root(drain, o!());
+    let logger = Logger::root(drain.fuse(), o!());
 
     debug!(logger, "debug"; "key" => "value");
     info!(logger, "info"; "key" => "value");
     warn!(logger, "warn"; "key" => "value");
     error!(logger, "error"; "key" => "value");
 
-    // wait a while to let the sender thread to finish its work
-    thread::sleep(Duration::from_secs(10));
+    // flush all log messages
+    fhandle.flush().wait().unwrap();
 }