+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Not released yet
 
 * Fix duplicate keys in the JSON message
+* Add API for flushing the drain
 
 ## v0.1.1 (2018-05-02)